+++ CHANGES.rst
@@ -6,6 +6,8 @@ Changelog
 
 1.0a1 (unreleased)
 ------------------
+- Exported Geolocation object info.
+  [daniele]
 
 - Initial release.
   [eikichi18]